refactor(ptx): remove hard-coded types.d.ts
