+++ libs/ptx/src/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './category';
+export * from './shared-types';
