+++ libs/ptx/src/restaurants.ts
@@ -1,12 +1,60 @@
 import { apiGet } from './lib/api';
+import { PTXCityMap } from './lib/category';
+import { City, Picture, Position } from './lib/shared-types';
 import { constructRestaurantsSearch } from './lib/utils';
-import { PTXCityMap } from './category';
-import { PTX } from './types';
+
+export interface Restaurant {
+  ID: string;
+  Name: string;
+  Description: string;
+  Address: string;
+  ZipCode?: string;
+  Phone: string;
+  OpenTime?: string;
+  WebsiteUrl?: string;
+  Picture: Picture;
+  Position: Position;
+  Class?: RestaurantClass;
+  SrcUpdateTime: string;
+  UpdateTime: string;
+  City?: string;
+  ParkingInfo?: string;
+  MapUrl?: string;
+}
+
+export type RestaurantClass =
+  | '中式美食'
+  | '伴手禮'
+  | '其他'
+  | '地方特產'
+  | '夜市小吃'
+  | '甜點冰品'
+  | '異國料理'
+  | '素食';
+
+export interface RestaurantCard {
+  ID: string;
+  Name: string;
+  City: string;
+  Address: string;
+  OpenTime?: string;
+  Phone?: string;
+  Picture: Picture;
+}
+
+export interface RestaurantRemark {
+  ID: string;
+  Name: string;
+  Description: string;
+  City: string;
+  Address: string;
+  Picture: Picture;
+}
 
 export const getRestaurantById = async (
   id: string,
-): Promise<PTX.Restaurant | undefined> => {
-  const result = await apiGet<PTX.Restaurant[]>('Tourism/Restaurant', {
+): Promise<Restaurant | undefined> => {
+  const result = await apiGet<Restaurant[]>('Tourism/Restaurant', {
     $top: '1',
     $filter: `ID eq '${id}'`,
   });
@@ -16,7 +64,7 @@ export const getRestaurantById = async (
 
 export const getRestaurantCards = async (
   count = 30,
-): Promise<PTX.RestaurantCard[]> =>
+): Promise<RestaurantCard[]> =>
   apiGet('Tourism/Restaurant', {
     $top: count.toString(),
     $select: 'ID,Name,City,Address,OpenTime,Phone,Picture',
@@ -25,9 +73,9 @@ export const getRestaurantCards = async (
   });
 
 export const getRestaurantCardsByCity = async (
-  city: PTX.City,
+  city: City,
   count = 30,
-): Promise<PTX.RestaurantCard[]> =>
+): Promise<RestaurantCard[]> =>
   apiGet(`Tourism/Restaurant/${PTXCityMap[city]}`, {
     $top: count.toString(),
     $select: 'ID,Name,City,Address,OpenTime,Phone,Picture',
@@ -37,7 +85,7 @@ export const getRestaurantCardsByCity = async (
 
 export const getRestaurantWithRemarks = async (
   count = 30,
-): Promise<PTX.RestaurantRemark[]> =>
+): Promise<RestaurantRemark[]> =>
   apiGet('Tourism/Restaurant', {
     $top: count.toString(),
     $select: 'ID,Name,Description,City,Address,Picture',
@@ -46,9 +94,9 @@ export const getRestaurantWithRemarks = async (
   });
 
 export const getRestaurantWithRemarksByCity = async (
-  city: PTX.City,
+  city: City,
   count = 30,
-): Promise<PTX.RestaurantRemark[]> =>
+): Promise<RestaurantRemark[]> =>
   apiGet(`Tourism/Restaurant/${PTXCityMap[city]}`, {
     $top: count.toString(),
     $select: 'ID,Name,Description,City,Address,Picture',
@@ -59,7 +107,7 @@ export const getRestaurantWithRemarksByCity = async (
 export const searchRestaurantsByKeyword = async (
   keyword: string,
   count = 30,
-): Promise<PTX.RestaurantCard[]> =>
+): Promise<RestaurantCard[]> =>
   apiGet('Tourism/Restaurant', {
     $top: count.toString(),
     $select: 'ID,Name,City,Address,OpenTime,Phone,Picture',
